+++ cli/src/command/build.rs
@@ -68,10 +68,9 @@ pub fn prepare_command_build<'a>(
     config_args: &'a mut Vec<&'a OsStr>,
 ) -> Result<Vec<&'a OsStr>> {
     let mut args: Vec<&OsStr> = vec![];
-    
+
     args.append(config_args);
     args.push(OsStr::new("build"));
-
 
     if args_matches.get_flag("BUILD_ARG") {
         args.push(OsStr::new("--build-arg"));