+++ cli/src/utils/system_tests.rs
@@ -3,8 +3,8 @@ mod tests {
     use std::ffi::OsStr;
 
     use crate::{
-        utils::system::{builder, System},
         parser::config::ComposeItem,
+        utils::system::{builder, System},
     };
 
     #[test]